+++ dbm-ui/backend/flow/engine/controller/redis.py
@@ -66,6 +66,9 @@ from backend.flow.engine.bamboo.scene.redis.redis_storages_client_conns_kill imp
     RedisStoragesClientConnsKillSceneFlow,
 )
 from backend.flow.engine.bamboo.scene.redis.redis_twemproxy_cluster_apply_flow import RedisClusterApplyFlow
+from backend.flow.engine.bamboo.scene.redis.revoke.redis_rollback_exercise_revoke_flow import (
+    RedisRollbackExerciseRevokeFlow,
+)
 from backend.flow.engine.bamboo.scene.redis.single_proxy_shutdown import SingleProxyShutdownFlow
 from backend.flow.engine.bamboo.scene.redis.tendisplus_lightning_data import TendisPlusLightningData
 from backend.flow.engine.bamboo.scene.redis.validate.redis_keystat_validator import RedisKeyStatFlowValidator
@@ -74,6 +77,7 @@ from backend.flow.engine.bamboo.scene.redis.validate.redis_migrate_validator imp
     RedisSingleInsMigrateFlowValidator,
 )
 from backend.flow.engine.controller.base import BaseController
+from backend.flow.engine.revoke.base import revoke_with
 from backend.flow.engine.validate.base_validate import validates_with
 
 
@@ -476,6 +480,7 @@ class RedisController(BaseController):
         flow = FailoverDrillFlow(root_id=self.root_id, data=self.ticket_data)
         flow.failover_drill()
 
+    @revoke_with(RedisRollbackExerciseRevokeFlow)
     def redis_rollback_exercise(self):
         """
         redis 回档演练